♻️ refactor(web): retire the per-agent skills page
The profile's skills tab is now the only place an agent's skills live.
The standalone page's capabilities move in first: the detail sheet
becomes the real inspector (description, version, model invocation,
convert-to-manual, upgrade check, delete, file editing), and a single ＋
opens an install sheet with market browsing and manual GitHub/ZIP
install sharing one scope picker.

The old URLs redirect to the profile's skills tab so existing links and
bookmarks keep working.

Refs #884
