+++ web/src/lib/i18n/messages.ts
@@ -83,20 +83,8 @@ const en = {
   "profile.skills": "Skills",
   "profile.tools": "Tools",
   "profile.skillsDesc": "Everything this agent can call on, grouped by who owns it.",
-  "profile.skillMenuNew": "New skill",
-  "profile.skillMenuUpload": "Upload",
-  "profile.skillMenuMarket": "Install from market",
-  "profile.uploadSkillTitle": "Upload a skill",
-  "profile.uploadSkillDesc": "A .zip holding one folder with a SKILL.md at its root.",
-  "profile.uploadSkillTarget": "Installs to your profile, available to every agent you own.",
-  "profile.uploadSkillAction": "Upload",
-  "profile.skillUploaded": "Skill uploaded",
-  "profile.skillUploadFailed": "Failed to upload skill",
-  "profile.skillDetails": "Skill details",
-  "profile.skillFiles": "Files",
   "profile.skillNoFiles": "This skill has no readable files.",
   "profile.skillFilesUnavailable": "Files are unavailable for this skill right now.",
-  "profile.skillEditContent": "Edit content",
   "profile.openSkill": "Open skill details",
   "profile.skillsGroupMine": "Mine",
   "profile.skillsGroupSystem": "System",
@@ -1702,11 +1690,6 @@ const en = {
   "sessions.workspace.share": "Share",
   "sessions.workspace.rename": "Rename",
   "sessions.workspace.binaryFile": "Binary file",
-  "sessions.skillsList.noSkills": "No skills yet",
-  "sessions.skillsList.user": "User",
-  "sessions.skillsList.agent": "Agent",
-  "sessions.skillsList.system": "System",
-  "sessions.skillsList.project": "Project",
   "sessions.skillsList.userDesc": "Your personal skills, visible only to you",
   "sessions.skillsList.user_agentDesc": "Your skills for this agent only",
   "sessions.skillsList.system_agentDesc": "Shared with everyone who uses this agent",
@@ -1715,13 +1698,10 @@ const en = {
   "sessions.skillsList.noAgentSkills": "No agent-level skills installed.",
   "sessions.skillsList.noSystemSkills": "No system skills.",
   "sessions.skillsList.noProjectSkills": "No skills in the project directory.",
-  "sessions.skillsList.installedTab": "Installed",
   "sessions.skillsList.discoverTab": "Discover",
-  "sessions.skillsList.all": "All",
   "sessions.skillsList.market": "Marketplace",
   "sessions.skillsList.manualTab": "Manual",
   "sessions.skillsList.builtin": "Built-in",
-  "sessions.skillsList.scopeFilter": "Scope",
   "sessions.skillsList.uploadZip": "Upload ZIP",
   "sessions.skillsList.installGithub": "Install from GitHub",
   "sessions.skillsList.githubRepo": "Repository",
@@ -1741,7 +1721,6 @@ const en = {
   "sessions.skillsList.manual": "Manual",
   "sessions.skillsList.auto": "Auto",
   "sessions.skillsList.files": "Files",
-  "sessions.skillsList.scope": "Scope",
   "sessions.skillsList.status": "Status",
   "sessions.skillsList.versionLabel": "Version",
   "sessions.skillsList.versionHint": "Recorded tag, branch, or commit. Cleared if left blank.",
@@ -1761,9 +1740,6 @@ const en = {
     "Apply your edits and stop Reflect and Curator from maintaining this skill.",
   "sessions.skillsList.convertToManualWarning":
     "This cannot be undone. The skill will no longer be maintained automatically.",
-  "sessions.skillsList.loadError": "Skills could not be loaded",
-  "sessions.skillsList.loadErrorDesc": "Check the connection and retry this page.",
-  "sessions.skillsList.allLoaded": "All skills loaded",
   "sessions.skillsList.saved": "Settings saved",
   "sessions.skillsList.emptyFile": "This file is empty.",
   "sessions.skillsList.name": "Name",
@@ -2170,20 +2146,8 @@ const zh: Record<MessageKey, string> = {
   "profile.skills": "技能",
   "profile.tools": "工具",
   "profile.skillsDesc": "该智能体可以调用的全部技能，按归属分组。",
-  "profile.skillMenuNew": "新建技能",
-  "profile.skillMenuUpload": "上传",
-  "profile.skillMenuMarket": "从市场安装",
-  "profile.uploadSkillTitle": "上传技能",
-  "profile.uploadSkillDesc": "一个 .zip 压缩包，内含单个文件夹，根目录须有 SKILL.md。",
-  "profile.uploadSkillTarget": "安装到你的个人档案，你名下的所有智能体都可使用。",
-  "profile.uploadSkillAction": "上传",
-  "profile.skillUploaded": "技能已上传",
-  "profile.skillUploadFailed": "上传技能失败",
-  "profile.skillDetails": "技能详情",
-  "profile.skillFiles": "文件",
   "profile.skillNoFiles": "该技能没有可读文件。",
   "profile.skillFilesUnavailable": "暂时无法读取该技能的文件。",
-  "profile.skillEditContent": "编辑内容",
   "profile.openSkill": "查看技能详情",
   "profile.skillsGroupMine": "我的",
   "profile.skillsGroupSystem": "系统",
@@ -3757,11 +3721,6 @@ const zh: Record<MessageKey, string> = {
   "sessions.workspace.share": "分享",
   "sessions.workspace.rename": "重命名",
   "sessions.workspace.binaryFile": "二进制文件",
-  "sessions.skillsList.noSkills": "暂无技能",
-  "sessions.skillsList.user": "用户",
-  "sessions.skillsList.agent": "智能体",
-  "sessions.skillsList.system": "系统",
-  "sessions.skillsList.project": "项目",
   "sessions.skillsList.userDesc": "你的个人技能，仅你可见",
   "sessions.skillsList.user_agentDesc": "仅用于此智能体的个人技能",
   "sessions.skillsList.system_agentDesc": "此智能体的所有使用者共享",
@@ -3770,9 +3729,7 @@ const zh: Record<MessageKey, string> = {
   "sessions.skillsList.noAgentSkills": "暂无智能体级技能。",
   "sessions.skillsList.noSystemSkills": "暂无系统技能。",
   "sessions.skillsList.noProjectSkills": "项目目录中暂无技能。",
-  "sessions.skillsList.installedTab": "已安装",
   "sessions.skillsList.discoverTab": "发现",
-  "sessions.skillsList.all": "全部",
   "sessions.skillsList.installGithub": "从 GitHub 安装",
   "sessions.skillsList.githubRepo": "仓库",
   "sessions.skillsList.githubRepoPlaceholder": "owner/repo 或 github.com 链接",
@@ -3787,7 +3744,6 @@ const zh: Record<MessageKey, string> = {
   "sessions.skillsList.market": "市场",
   "sessions.skillsList.manualTab": "手动安装",
   "sessions.skillsList.builtin": "系统内置",
-  "sessions.skillsList.scopeFilter": "范围",
   "sessions.skillsList.uploadZip": "上传 ZIP",
   "sessions.skillsList.searchPlaceholder": "搜索技能…",
   "sessions.skillsList.stats": "{{total}} 项技能 · {{callable}} 项可自动调用 · {{readonly}} 项只读",
@@ -3795,7 +3751,6 @@ const zh: Record<MessageKey, string> = {
   "sessions.skillsList.manual": "手动",
   "sessions.skillsList.auto": "自动",
   "sessions.skillsList.files": "文件",
-  "sessions.skillsList.scope": "作用域",
   "sessions.skillsList.status": "状态",
   "sessions.skillsList.versionLabel": "版本",
   "sessions.skillsList.versionHint": "记录的 tag、branch 或 commit。留空则清除。",
@@ -3813,9 +3768,6 @@ const zh: Record<MessageKey, string> = {
   "sessions.skillsList.convertToManualHint":
     "保存本次编辑，并停止 Reflect 和 Curator 继续维护此技能。",
   "sessions.skillsList.convertToManualWarning": "此操作不可逆，之后将无法自动维护。",
-  "sessions.skillsList.loadError": "技能加载失败",
-  "sessions.skillsList.loadErrorDesc": "请检查连接后重试。",
-  "sessions.skillsList.allLoaded": "已加载全部技能",
   "sessions.skillsList.saved": "设置已保存",
   "sessions.skillsList.emptyFile": "此文件为空。",
   "sessions.skillsList.name": "名称",
