+++ web/src/lib/route-search.ts
@@ -12,16 +12,6 @@ export type ProfileTab = "overview" | "memory" | "skills" | "tools" | "config";
 
 const PROFILE_TABS = new Set<string>(["overview", "memory", "skills", "tools", "config"]);
 
-export interface SkillsSearch {
-  new?: boolean;
-  source?: "installed" | "market" | "manual";
-  fscope?: "project" | "user" | "agent" | "system";
-  sel?: string;
-}
-
-const SKILL_SOURCES = new Set(["installed", "market", "manual"]);
-const SKILL_SCOPES = new Set(["project", "user", "agent", "system"]);
-
 export interface ThreadsSearch {
   /**
    * Which home the thread list is scoped to: absent means every home, "agent"
@@ -44,18 +34,5 @@ export function validateMemorySearch(search: Record<string, unknown>): MemorySea
   return {
     ...(search.knowledge === "removed" ? { knowledge: "removed" as const } : {}),
     ...(PROFILE_TABS.has(search.tab as string) ? { tab: search.tab as ProfileTab } : {}),
-  };
-}
-
-export function validateSkillsSearch(search: Record<string, unknown>): SkillsSearch {
-  return {
-    new: search.new === true || search.new === "true",
-    source: SKILL_SOURCES.has(search.source as string)
-      ? (search.source as SkillsSearch["source"])
-      : undefined,
-    fscope: SKILL_SCOPES.has(search.fscope as string)
-      ? (search.fscope as SkillsSearch["fscope"])
-      : undefined,
-    sel: typeof search.sel === "string" ? search.sel : undefined,
   };
 }
