--- web/src/routes/_app/agents.$agentId/skills/$skillId.lazy.tsx
@@ -1,6 +1,0 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
-import { SkillEditPage } from "@/features/sessions/pages/SkillEditPage";
-
-export const Route = createLazyFileRoute("/_app/agents/$agentId/skills/$skillId")({
-  component: SkillEditPage,
-});
