+++ web/src/routes/_app/agents.$agentId/skills/index.tsx
@@ -1,6 +1,15 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { validateSkillsSearch } from "@/lib/route-search";
+import { createFileRoute, redirect } from "@tanstack/react-router";
 
+// An agent's skills live on its profile; the standalone skills page is gone.
+// The old URL stays valid as a redirect so existing links and bookmarks land on
+// the profile's skills tab.
 export const Route = createFileRoute("/_app/agents/$agentId/skills/")({
-  validateSearch: validateSkillsSearch,
+  beforeLoad: ({ params: { agentId } }) => {
+    throw redirect({
+      to: "/agents/$agentId/profile",
+      params: { agentId },
+      search: { tab: "skills" as const },
+      replace: true,
+    });
+  },
 });
